# chore: Optimize SQL performance for tables with large number of columns