--- a/querybook/server/logic/tag.py
+++ b/querybook/server/logic/tag.py
@@ -30,6 +30,31 @@ def get_tags_by_column_id(column_id: int, session=None):
 
 
 @with_session
+def get_tags_by_column_ids(column_ids: int, session=None):
+    """Get all tags for multiple columns at once, organized by column_id"""
+    if not column_ids:
+        return {}
+
+    # Query all tags for the given column IDs in a single query
+    tag_results = (
+        session.query(Tag, TagItem.column_id)
+        .join(TagItem)
+        .filter(TagItem.column_id.in_(column_ids))
+        .order_by(Tag.count.desc())
+        .all()
+    )
+
+    # Organize results by column_id
+    tags_by_column = {}
+    for tag, column_id in tag_results:
+        if column_id not in tags_by_column:
+            tags_by_column[column_id] = []
+        tags_by_column[column_id].append(tag)
+
+    return tags_by_column
+
+
+@with_session
 def get_tags_by_keyword(keyword, limit=10, session=None):
     return (
         session.query(Tag)