This PR introduces performance optimizations for tables with many columns by reducing database load and query count:

### Key Changes:
- **Batch loading**: Implemented consolidated batch queries for columns, tags, and data element associations instead of individual queries
- **Optimized `get_table_schemas_by_names`**: Completely rewritten to:
  - Fetch tables, columns, and tags in separate efficient queries
  - Use SQLAlchemy's `load_only` and `joinedload` to minimize data transfer
  - Process results in memory rather than making many small queries
- **Optimized `get_detailed_columns_dict_by_table_id`**: Reimplemented to:
  - Retrieve all column data in a single query
  - Batch load tags and data element associations for all columns at once
  - Assemble column data in memory instead of making separate queries per column
- **Model improvements**:
  - Added `lazy="joined"` optimization for column statistics as `stats` will be retrieved by default
  - Added `to_dict()` method for DataTableColumnStatistics to support JSON serialization, which will make `flask.jsonify` more efficient.

### Performance Improvements:
- **Single table with 581 columns**:
  - `/ds/table/<table-id>/`: ~10s → <1s
  - `/ds/table/<table-id>/detailed_column/`: 5-6s → <1s
- **Multiple tables (10 tables including large table)**:
  - Schema retrieval: 9-10s → <2s